stop words properties definition replaced by StopWordsMixin
