+++ sumy/summarizers/lsa.py
@@ -16,21 +16,13 @@ try:
     from numpy.linalg import svd as singular_value_decomposition
 except ImportError:
     singular_value_decomposition = None
+
 from ._summarizer import AbstractSummarizer
+from ._mixins import StopWordsMixin
 
-
-class LsaSummarizer(AbstractSummarizer):
+class LsaSummarizer(AbstractSummarizer, StopWordsMixin):
     MIN_DIMENSIONS = 3
     REDUCTION_RATIO = 1/1
-    _stop_words = frozenset()
-
-    @property
-    def stop_words(self):
-        return self._stop_words
-
-    @stop_words.setter
-    def stop_words(self, words):
-        self._stop_words = frozenset(map(self.normalize_word, words))
 
     def __call__(self, document, sentences_count):
         self._ensure_dependecies_installed()
